feat: add relation between user and tasks

+++ src/tasks/tasks.service.ts
@@ -1,5 +1,6 @@
 import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/users/user.entity';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { GetTasksFilterDto } from './dto/get-tasks-filter.dto';
 import { TaskStatus } from './task-status.enum';
@@ -12,11 +13,13 @@ export class TasksService {
     @InjectRepository(TaskRepository) private taskRepository: TaskRepository,
   ) {}
 
-  async getTasks(searchFilter: GetTasksFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(searchFilter);
+  async getTasks(searchFilter: GetTasksFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(searchFilter, user);
   }
-  async getATask(id: number): Promise<Task> {
-    const record = await this.taskRepository.findOne({ where: { id } });
+  async getATask(id: number, user: User): Promise<Task> {
+    const record = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
 
     if (!record) {
       throw new NotFoundException(`task with id ${id} not found`);
@@ -24,16 +27,20 @@ export class TasksService {
     return record;
   }
 
-  async createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.createTask(createTaskDto);
+  async createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.createTask(createTaskDto, user);
   }
-  async deleteATask(id: number): Promise<void> {
-    const isDeleted = await this.taskRepository.delete(id);
+  async deleteATask(id: number, user: User): Promise<void> {
+    const isDeleted = await this.taskRepository.delete({ id, userId: user.id });
     if (isDeleted.affected === 0)
       throw new NotFoundException(`task with id ${id} not found`);
   }
-  async updateStatusOfATask(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getATask(id);
+  async updateStatusOfATask(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getATask(id, user);
     task.status = status;
     await task.save();
     return task;
